refactor: 나의 강좌리스트 api호출을 한번 하도록 개선
